fix deleteFile issue

+++ src/controllers/fileControllers/deleteFile.js
@@ -4,9 +4,13 @@ const Pdf = require("../../models/pdfModel");
 async function deleteFile(req,res){
   try{
      const file = await Pdf.findOne({
-         userId: req.user.id,
          _id: req.params.id
        });
+      if(file.userId.equals(req.user.id)){
+         return res.status(404).json({
+            status: "no file exist with this id to be deleted",
+          });  
+      }
        if(file){
         await file.delete();
         res.status(204).json({
